+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The **need for configuration updates** is **marked bold**
 ## Added
 
 - Updated user guide with data exchange forwarding feature ([#1214](https://github.com/eclipse-tractusx/puris/pull/1214))
+- Updated frontend to allow data exchange requests for demand-based effects ([#1214](https://github.com/eclipse-tractusx/puris/pull/1214))
 
 ### Changed